requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
discord.py==2.2.3
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
multidict==6.0.4
mutagen==1.46.0
pycparser==2.21
pycryptodomex==3.18.0
PyNaCl==1.5.0
redis==4.5.5
requests==2.31.0
six==1.16.0
sniffio==1.3.0
spotipy==2.23.0
urllib3==2.0.3
websockets==11.0.3
yarl==1.9.2
youtube-search-python==1.6.6
yt-dlp==2023.3.4
